+++ docs/changes.rst
@@ -9,6 +9,10 @@ To be released.
 - Fixed a bug that :class:`~sassutils.wsgi.SassMiddleware` yielded
   :class:`str` instead of :class:`bytes` on Python 3.
 - Fixed several Unicode-related bugs on Windows.
+- Fixed a bug that :func:`~sassutils.builder.build_directory()`,
+  :class:`~sassutils.wsgi.SassMiddleware`, and
+  :class:`~sassutils.distutils.build_sass` don't recursively build
+  subdirectories.
 
 
 Version 0.5.0